+++ spring-cloud-tencent-rpc-enhancement/src/main/java/com/tencent/cloud/rpc/enhancement/plugin/PluginOrderConstant.java
@@ -58,5 +58,21 @@ public class PluginOrderConstant {
 		 * {@link com.tencent.cloud.metadata.core.EncodeTransferMedataRestTemplateEnhancedPlugin}.
 		 */
 		public static final int CONSUMER_TRANSFER_METADATA_PLUGIN_ORDER = Ordered.HIGHEST_PRECEDENCE + 10;
+
+		/**
+		 * order for
+		 * {@link com.tencent.cloud.plugin.trace.TraceMetadataEnhancedPlugin}.
+		 */
+		public static final int CONSUMER_TRACE_METADATA_PLUGIN_ORDER = CONSUMER_TRANSFER_METADATA_PLUGIN_ORDER - 1;
 	}
+
+	public static class ServerPluginOrder {
+
+		/**
+		 * order for
+		 * {@link com.tencent.cloud.plugin.trace.TraceServerMetadataEnhancedPlugin}.
+		 */
+		public static final int PROVIDER_TRACE_METADATA_PLUGIN_ORDER = Ordered.HIGHEST_PRECEDENCE + 1;
+	}
+
 }